src/main/java/com/kakaopay/coffeeorder/service/MenuService.java
package com.kakaopay.coffeeorder.service;

import java.util.List;

import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.kakaopay.coffeeorder.domain.Menu;
import com.kakaopay.coffeeorder.repository.MenuRepository;

import lombok.RequiredArgsConstructor;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class MenuService {

	private final MenuRepository menuRepository;

	public List<Menu> findAllMenus() {
		return menuRepository.findAll();
	}
}
